Channel for PRs and new struct
Created necessary fields inside the estimator state struct to be able to receive new providers from the putProviderMethod using a channel. The providers are saved inside a map, a field of the estimator state struct.
Created a new type to do that cidAndProvider struct and this is sent over the channel.
When the channel receives a doneProviderChannel message it calls the saveProvidersToFile method insdie the record_providers file.